requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
tqdm==4.66.4
urllib3==2.2.2
yarl==1.9.4
